Move test files next to the class being tested
