+++ charts/BinningStrategies.test.ts
