+++ charts/ExploreView.test.tsx
@@ -4,16 +4,16 @@ import * as React from "react"
 import { shallow, mount, ReactWrapper } from "enzyme"
 import { observe } from "mobx"
 
-import { ExploreView } from "../ExploreView"
-import { Bounds } from "../Bounds"
-import { ChartView } from "../ChartView"
-import { ChartType } from "../ChartType"
-import { LineChart } from "../LineChart"
-import { StackedArea } from "../StackedArea"
-import { StackedBarChart } from "../StackedBarChart"
-import { DiscreteBarChart } from "../DiscreteBarChart"
-import { SlopeChart } from "../SlopeChart"
-import { ChoroplethMap } from "../ChoroplethMap"
+import { ExploreView } from "./ExploreView"
+import { Bounds } from "./Bounds"
+import { ChartView } from "./ChartView"
+import { ChartType } from "./ChartType"
+import { LineChart } from "./LineChart"
+import { StackedArea } from "./StackedArea"
+import { StackedBarChart } from "./StackedBarChart"
+import { DiscreteBarChart } from "./DiscreteBarChart"
+import { SlopeChart } from "./SlopeChart"
+import { ChoroplethMap } from "./ChoroplethMap"
 import { RootStore } from "charts/Store"
 import { ExploreModel } from "charts/ExploreModel"
