+++ explorer/client/covidDataExplorer/CovidConstants.ts
