+++ explorer/client/covidDataExplorer/CovidExplorerTable.ts
