+++ site/server/views/ExplorePage.test.tsx
@@ -3,9 +3,9 @@
 import * as React from "react"
 import { shallow } from "enzyme"
 
-import { ExplorePage } from "../ExplorePage"
-import { SiteHeader } from "../SiteHeader"
-import { SiteFooter } from "../SiteFooter"
+import { ExplorePage } from "./ExplorePage"
+import { SiteHeader } from "./SiteHeader"
+import { SiteFooter } from "./SiteFooter"
 
 describe(ExplorePage, () => {
     it("renders a site header", () => {
